requirements.txt
#
# Please do not manually update this file since the requirements are managed
# by pipenv through Pipfile and Pipfile.lock .
#
# This file is created and managed automatically by the template and it is left
# here only for backwards compatibility reasons with python's ecosystem.
#
# Please use Pipfile to update the requirements.
#
azure-mgmt-security~=1.0.0
azure-identity~=1.9.0
azure-cli-core~=2.35.0
azure-mgmt-resourcegraph~=8.0.0
azure-mgmt-resource~=21.0.0
schema~=0.7.5
pandas~=1.3.5
cachetools~=4.2.4
opnieuw~=1.1.0